projectsCard, retrieving correct information in separeted individual cards
