+++ vin-port/src/app/page.tsx
@@ -1,5 +1,6 @@
 import AboutSection from "@/components/AboutSection";
 import Header from "@/components/Header";
+import ProjectsSection from "@/components/ProjectsSection";
 import WelcomeSection from "@/components/WelcomeSection";
 
 export default function Home() {
@@ -8,7 +9,7 @@ export default function Home() {
         <Header />
         <WelcomeSection />
         <AboutSection />  
-          
+        <ProjectsSection />
     </main>
   );
 }
